fix(react): address PR review feedback
- Move `HiddenActions` definition to `SurveyFormBuilder/types.ts` so the
  module graph stays one-directional. `BaseQuestion/types.ts` now imports
  it from there (it already depends on the same file for `QuestionType`)
  and re-exports it for backward-compatible access via BaseQuestion.
- Tighten the Storybook description for the unrestricted free-text
  question: the default menu also includes the `Question type` entry.
